backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pydantic-settings>=2.2.1
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
email-validator>=2.1.1
requests>=2.32.3
python-jose[cryptography]==3.3.0
websockets>=12.0
openai
httpx>=0.27.0
